fix cho nhập bao nhiêu số thập phân cũng được phần số lượng

+++ server/migrations/2025_07_09_000000_change_quantity_fees_to_decimal_in_orders_table.php
@@ -14,7 +14,7 @@ class ChangeQuantityFeesToDecimalInOrdersTable extends Migration
     public function up()
     {
         Schema::table('orders', function (Blueprint $table) {
-            $table->decimal('quantity_fees', 10, 2)->default(1)->change();
+            $table->double('quantity_fees')->default(1)->change();
         });
     }
